% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_hyptest.R
\name{complex_hyptest}
\alias{complex_hyptest}
\title{Performs a hypothesis test on your regression coefficients (Can also do hypothesis on non-regression related problems).
This will tell inform you if placing constraints on your Beta coefficients is reasonable given the covariance structure.
The complex_hyptest function can perform the basics equality vs. inequality (EI) and inequality vs. unrestricted (IU) tests
like hyptest but adds a level of complexity to the problem. Unlike hyptest, complex_hyptest tacks on
another set of equality restrictions (Ax=b) into the null hypothesis. In both the EI and
IU cases, the alternative associated with these new restrictions is unrestricted.
A small p-value will indicate that the constraints you want to impose on your regression are unreasonable and do not respect the covariance structure of your data
If all your constraints are equality or inequality, don't use this function; the hyptest function in this package will work better and faster.
DISCLAIMER: Like all functions in this package, this test assumes normality in both the null and alternative distributions and uses the estimated covariance to construct both.
In the context of regression, these assumptions will hold a large proportion of the time, but it is important to check prior to using this function and the package in general.}
\usage{
complex_hyptest(cov, mean_hat, G, h, A, b, type, input_draws = 10000)
}
\arguments{
\item{G}{The hypothesized quadratic inequality constraint matrix such that: GB >= h}

\item{h}{The hypothesized quadratic inequality constraint vector such that: GB >= h}

\item{A}{The hypothesized quadratic equality constraint matrix such that: AB = b}

\item{b}{The hypothesized quadratic equality constraint vector such that: AB = b}

\item{cov:}{estimated covariance matrix}

\item{mean_hat:}{estimated mean}

\item{type:}{To do the IU test, input "IU". To do the EI test, input "EI"}
}
\value{
Graphs the distribution of the test statistic and shows you how likely your unrestricted Beta estimate is given the restrictions you entered. REMEMBER: THIS IS A ONE SIDED TEST!!!
}
\description{
Performs a hypothesis test on your regression coefficients (Can also do hypothesis on non-regression related problems).
This will tell inform you if placing constraints on your Beta coefficients is reasonable given the covariance structure.
The complex_hyptest function can perform the basics equality vs. inequality (EI) and inequality vs. unrestricted (IU) tests
like hyptest but adds a level of complexity to the problem. Unlike hyptest, complex_hyptest tacks on
another set of equality restrictions (Ax=b) into the null hypothesis. In both the EI and
IU cases, the alternative associated with these new restrictions is unrestricted.
A small p-value will indicate that the constraints you want to impose on your regression are unreasonable and do not respect the covariance structure of your data
If all your constraints are equality or inequality, don't use this function; the hyptest function in this package will work better and faster.
DISCLAIMER: Like all functions in this package, this test assumes normality in both the null and alternative distributions and uses the estimated covariance to construct both.
In the context of regression, these assumptions will hold a large proportion of the time, but it is important to check prior to using this function and the package in general.
}
