% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyptest.R
\name{hyptest}
\alias{hyptest}
\title{Performs the IU and EI hypothesis tests which are especially useful in the context of constrained regression.
IU: Shorthand for null hypothesis of inequality constraints vs. an unrestricted alternative.
EI: Shorthand for equality constraints null hypothesis vs. inequality constraints alternative (greater than the null).
The IU hypothesis test in this function assumes that all null constraints that you enter are inequality constraints.
The EI hypothesis test in this function assumes that all null constraints that you enter are equality constraints.
For a more complex analysis that can take non-diagonal constraint matrices and employ inequality and equality constraints at the same time, use complex_hyptest.
DISCLAIMER: Like all functions in this package, both tests assumes normality in both the null and alternative distributions and uses the estimated covariance to construct both.
In the context of regression, these assumptions will hold a large proportion of the time, but it is important to check prior to using this function and the package in general.}
\usage{
hyptest(cov, mean_hat, R, r, type, input_draws = 10000)
}
\arguments{
\item{cov_matrix:}{estimated covariance matrix}

\item{mean_hat:}{estimated mean}

\item{G:}{Must be DIAGONAL The hypothesized quadratic constraint matrix such that: Gx >= h or Gx = h under the null (both tests are done here)}

\item{h:}{The hypothesized quadratic constraint vector such that: Gx >= h or Gx = h under the null (both tests are done here)}

\item{type:}{To do the IU test, input "IU". To do the EI test, input "EI"}
}
\value{
Graphs the CDF of the EI and IU distributions and the p-value of the respective hypothesis test (remember this is always a one-sided test!)
}
\description{
Performs the IU and EI hypothesis tests which are especially useful in the context of constrained regression.
IU: Shorthand for null hypothesis of inequality constraints vs. an unrestricted alternative.
EI: Shorthand for equality constraints null hypothesis vs. inequality constraints alternative (greater than the null).
The IU hypothesis test in this function assumes that all null constraints that you enter are inequality constraints.
The EI hypothesis test in this function assumes that all null constraints that you enter are equality constraints.
For a more complex analysis that can take non-diagonal constraint matrices and employ inequality and equality constraints at the same time, use complex_hyptest.
DISCLAIMER: Like all functions in this package, both tests assumes normality in both the null and alternative distributions and uses the estimated covariance to construct both.
In the context of regression, these assumptions will hold a large proportion of the time, but it is important to check prior to using this function and the package in general.
}
