% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_power.R
\name{mult_power}
\alias{mult_power}
\title{Simple power calculator that can easily applied to constrained regression problems.
This function will determine the power of an alternative vs. a null hypothesis given constraints, a covariance estimate, and alpha level.
DISCALIMER: Like all functions in this package, this power calculation assumes normality in both the null and alternative distributions and uses the estimated covariance to construct both.
In the context of regression, these assumptions will hold a large proportion of the time, but it is important to check prior to using this function and the package in general.}
\usage{
mult_power(cov, true_mu, G, h, A, b, type, alpha = 0.05)
}
\arguments{
\item{G}{The hypothesized quadratic inequality constraint matrix such that: GB >= h}

\item{h}{The hypothesized quadratic inequality constraint vector such that: GB >= h}

\item{A}{The hypothesized quadratic equality constraint matrix such that: AB = b}

\item{b}{The hypothesized quadratic equality constraint vector such that: AB = b}

\item{cov:}{The estimated covariance matrix}

\item{true_mu:}{The mean you wish to test against the null constraints. (Test your estimated mean - power is another metric to see how badly the constraints are violated)}

\item{type:}{To do the IU test, input "IU". To do the EI test, input "EI"}

\item{alpha:}{The parameter that dictates the level of signifigance you wish to test the power at. It will default to .05 if nothing else is entered.}
}
\value{
The power of true_mu vs. mu_hat (least favorable mean given the constraints)
}
\description{
Simple power calculator that can easily applied to constrained regression problems.
This function will determine the power of an alternative vs. a null hypothesis given constraints, a covariance estimate, and alpha level.
DISCALIMER: Like all functions in this package, this power calculation assumes normality in both the null and alternative distributions and uses the estimated covariance to construct both.
In the context of regression, these assumptions will hold a large proportion of the time, but it is important to check prior to using this function and the package in general.
}
